copilot/requirements.txt
uvicorn
dapr
python-dotenv
fastapi
openai
azure-identity
azure-messaging-webpubsubservice